Add Decodable & Encodable interfaces
