+++ Sources/SwiftAPI/Explorations/Decoding/RequestSingleValueDecodingContainer.swift
@@ -1,0 +1,67 @@
+//struct RequestSingleValueDecodingContainer: SingleValueDecodingContainer {
+//    var codingPath: [CodingKey]
+//
+//    func decodeNil() -> Bool {
+//        <#code#>
+//    }
+//
+//    func decode(_ type: Bool.Type) throws -> Bool {
+//        <#code#>
+//    }
+//
+//    func decode(_ type: String.Type) throws -> String {
+//        <#code#>
+//    }
+//
+//    func decode(_ type: Double.Type) throws -> Double {
+//        <#code#>
+//    }
+//
+//    func decode(_ type: Float.Type) throws -> Float {
+//        <#code#>
+//    }
+//
+//    func decode(_ type: Int.Type) throws -> Int {
+//        <#code#>
+//    }
+//
+//    func decode(_ type: Int8.Type) throws -> Int8 {
+//        <#code#>
+//    }
+//
+//    func decode(_ type: Int16.Type) throws -> Int16 {
+//        <#code#>
+//    }
+//
+//    func decode(_ type: Int32.Type) throws -> Int32 {
+//        <#code#>
+//    }
+//
+//    func decode(_ type: Int64.Type) throws -> Int64 {
+//        <#code#>
+//    }
+//
+//    func decode(_ type: UInt.Type) throws -> UInt {
+//        <#code#>
+//    }
+//
+//    func decode(_ type: UInt8.Type) throws -> UInt8 {
+//        <#code#>
+//    }
+//
+//    func decode(_ type: UInt16.Type) throws -> UInt16 {
+//        <#code#>
+//    }
+//
+//    func decode(_ type: UInt32.Type) throws -> UInt32 {
+//        <#code#>
+//    }
+//
+//    func decode(_ type: UInt64.Type) throws -> UInt64 {
+//        <#code#>
+//    }
+//
+//    func decode<T>(_ type: T.Type) throws -> T where T : Decodable {
+//        <#code#>
+//    }
+//}
