--- Sources/SwiftAPI/HTTPMethod.swift
@@ -1,3 +1,0 @@
-public enum HTTPMethod: String {
-    case post, put, get, delete, patch
-}
