+++ Sources/SwiftAPI/Methods/Method.swift
@@ -1,0 +1,11 @@
+public enum HTTPMethod: String {
+    case post, put, get, delete, patch
+}
+
+public class Method<Req, Res: Codable> {
+    public var wrappedValue: Endpoint<Req, Res>
+
+    init(_ method: HTTPMethod, _ basePath: String) {
+        self.wrappedValue = Endpoint<Req, Res>(method: method, basePath: basePath)
+    }
+}
