+++ Sources/SwiftAPI/Parameters/Header.swift
@@ -2,7 +2,7 @@ protocol AnyHeader {
     var value: String { get }
 }
 
-@propertyWrapper struct _Header: AnyHeader {
+@propertyWrapper struct Header: AnyHeader {
     var wrappedValue: String
     var value: String { wrappedValue }
 }
