+++ Sources/SwiftAPI/Parameters/Path.swift
@@ -2,7 +2,7 @@ protocol AnyPath {
     var value: String { get }
 }
 
-@propertyWrapper struct _Path: AnyPath {
+@propertyWrapper struct Path: AnyPath {
     var wrappedValue: String
     var value: String { wrappedValue }
 }
