+++ Sources/SwiftAPI/SwiftAPIError.swift
@@ -1,0 +1,4 @@
+struct SwiftAPIError: Error, CustomStringConvertible {
+    let message: String
+    var description: String { "[SwiftAPIError] \(self.message)" }
+}
