+++ Tests/SwiftAPITests/TestAPI.swift
@@ -1,16 +1,16 @@
 @testable import SwiftAPI
 
 struct Test {
-    @POST("v1/accounts/:userID:/transfer")
+    @POST("v1/accounts/{userID}/transfer")
     var test: Endpoint<TestReqDTO, TestResDTO>
 }
 
 struct TestReqDTO: RequestEncodable {
-    @_Path   var userID: String
-    @_Query  var number: Int
-    @_Query  var someThings: [String]
-    @_Header var value: String
-    @_Body   var obj: TestObj
+    @Path   var userID: String
+    @Query  var number: Int
+    @Query  var someThings: [String]
+    @Header var value: String
+    @Body   var obj: TestObj
 }
 
 struct TestResDTO: Codable {
